feat(timetable/querystring): 훅 구현/ timetable 쿼리스트링 적용
- 커스텀 훅 구현
- timetable 적용

Resolve: #167
